% Generated by roxygen2 (4.0.1): do not edit by hand
\name{animate_sm}
\alias{animate_sm}
\title{Make an animation of the model output in relation to the salt marsh community.}
\usage{
animate_sm()
}
\value{
a GIF animated file
}
\description{
It requires ImageMagick installed in the system to work.
}
\examples{
## Not run animate_sm()
}
\keyword{animation}

