\name{animate_ss}
\alias{animate_ss}
\title{Make an animation of the model output in relation to the salt steppe community.}
\usage{
  animate_ss()
}
\arguments{
  \item{inputParameter1}{A description of the input
  parameter \code{inputParameter1}}

  \item{inputParameter2}{A description of the input
  parameter \code{inputParameter2}}
}
\value{
  a GIF animated file
}
\description{
  It requires ImageMagick installed in the system to work.
}
\examples{
R code here showing how your function works
}
\keyword{animation}

