% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mrgf}
\alias{mrgf}
\title{A function to perform Multiple Resolution Goodness of Fit.}
\usage{
mrgf(a = "", b = "", w1 = 1, w2 = 113, year = "", k = 0)
}
\arguments{
\item{a}{a reference map}

\item{b}{b simulated map}

\item{w1}{w1 initial window size}

\item{w2}{w2 final window size}

\item{year}{year validation year}

\item{k}{k parameter for weighting Ft with lower/larger windows resolutions}
}
\value{
Fw vector of fits for each window size

Ft weighted overall fit
}
\description{
Returns the results of a Multiple Resolution Goodness of Fit after the modified method of Kuhnert et al. 2005, originally by Costanza 1989.
}
\examples{
## Not run mrgf(a='y1992_rs.asc',b='y1992_mod.asc',w1=1,w2=113,year='1992',k=0)
}
\keyword{Fit}
\keyword{Goodness}
\keyword{Multiple}
\keyword{Resolution}

