% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neigh_cell}
\alias{neigh_cell}
\title{A function to identify neighboring cells.}
\usage{
neigh_cell(cell)
}
\arguments{
\item{cell}{cell number from which to compute neighbor cells}
}
\value{
returns cell numbers of neighboring pixels to the reference one
}
\description{
Returns cell numbers of neighboring pixels to a reference cell
}
\examples{
## Not run neigh_cell(n)
}
\keyword{neighborhood}

