\name{plot_maps}
\alias{plot_maps}
\title{Plot abundance maps of plant communities.}
\usage{
  plot_maps(year = 2008)
}
\arguments{
  \item{year}{year to plot (from 1984 to 2008)}
}
\value{
  by default plots the final plant communities map (year =
  2008).
}
\description{
  A more detailed description of what the function is and
  how it works. It may be a paragraph that should not be
  separated by any spaces.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

