\docType{package}
\name{spdynmod}
\alias{spdynmod}
\alias{spdynmod-package}
\title{spdynmod}
\usage{
  spdynmod(t, init, parameters, nr, nc)
}
\arguments{
  \item{inputParameter1}{A description of the input
  parameter \code{inputParameter1}}

  \item{inputParameter2}{A description of the input
  parameter \code{inputParameter2}}
}
\value{
  output A description of the object the function outputs
}
\description{
  spdynmod

  A more detailed description of what the function is and
  how it works. It may be a paragraph that should not be
  separated by any spaces.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

