% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mrgf.R
\name{mrgf}
\alias{mrgf}
\title{A function to perform Multiple Resolution Goodness of Fit.}
\usage{
mrgf(year = "1992", w1 = 1, w2 = 113, k = 0)
}
\arguments{
\item{year}{year validation year}

\item{w1}{w1 initial window size}

\item{w2}{w2 final window size}

\item{k}{k parameter for weighting Ft with lower/larger windows resolutions}
}
\value{
Fw vector of fits for each window size

Ft weighted overall fit
}
\description{
Returns the results of a Multiple Resolution Goodness of Fit after the modified method of Kuhnert et al. 2005, originally by Costanza 1989.
}
\examples{
## Not run mrgf(year='1992',w1=1,w2=113,k=0)
}
\keyword{Fit}
\keyword{Goodness}
\keyword{Multiple}
\keyword{Resolution}

