% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_maps}
\alias{plot_maps}
\title{Plot abundance maps of plant communities.}
\usage{
plot_maps(year = 2008)
}
\arguments{
\item{year}{year to plot (from 1984 to 2008)}
}
\value{
by default plots the final plant communities map (year = 2008).
}
\description{
Plot abundance maps of plant communities in a given year.
}
\examples{
## Not run plot_maps()
}
\keyword{plot}

