\name{plot_maps}
\alias{plot_maps}
\title{Plot abundance maps of plant communities.}
\usage{
  plot_maps(i = 96)
}
\arguments{
  \item{i}{time step to plot (from 1 to 96)}

  \item{inputParameter2}{A description of the input
  parameter \code{inputParameter2}}
}
\value{
  by default plots the final plant communities map (i =
  96). There are 4 maps per year from 1984 to 2008.
}
\description{
  A more detailed description of what the function is and
  how it works. It may be a paragraph that should not be
  separated by any spaces.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

