% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{spdynmod}
\alias{spdynmod}
\alias{spdynmod-package}
\title{spdynmod}
\usage{
spdynmod(t, init, parameters, nr, nc)
}
\arguments{
\item{t}{time}

\item{init}{init}

\item{parameters}{model parameters}

\item{nr}{number of raster map rows}

\item{nc}{number of raster map columns}
}
\value{
Model solver
}
\description{
spdynmod

The wetland model function.
}
\examples{
## Not run spdynmod()
}
\keyword{solver}

